fix : 디렉토리 네임  등  Auto 들어가 있던 문구 삭제
