--- src/main/kotlin/autolotto/AutoLottoApplication.kt
@@ -1,21 +1,0 @@
-package autolotto
-
-import autolotto.configuration.AutoLottoConfiguration
-import autolotto.controller.AutoLottoController
-import autolotto.view.InputView
-import autolotto.view.OutPutView
-
-private val autoLottoController: AutoLottoController =
-    AutoLottoConfiguration().autoLottoController()
-
-fun main() {
-    val amount = InputView.getLottoPurchaseAmount()
-    InputView.printLottoGameCount(amount.lottoGameCount)
-
-    val createLottoInfo = autoLottoController.start(amount.lottoGameCount)
-    OutPutView.printLottoInfo(createLottoInfo)
-
-    val winnerNumbers = InputView.getWinningNumber()
-    val resultResponse = autoLottoController.getResult(winnerNumbers)
-    OutPutView.printLottoResults(resultResponse, amount.amount)
-}
