--- src/main/kotlin/autolotto/configuration/AutoLottoConfiguration.kt
@@ -1,19 +1,0 @@
-package autolotto.configuration
-
-import autolotto.controller.AutoLottoController
-import autolotto.repository.LottoRepository
-import autolotto.service.LottoService
-
-class AutoLottoConfiguration {
-    fun autoLottoController(): AutoLottoController {
-        return AutoLottoController(lottoService())
-    }
-
-    fun lottoService(): LottoService {
-        return LottoService(lottoRepository())
-    }
-
-    fun lottoRepository(): LottoRepository {
-        return LottoRepository()
-    }
-}
